Split fan checks out of temperature.cpp
